requirements.txt
cachetools==5.5.2
certifi==2025.6.15
charset-normalizer==3.4.2
google-api-core==2.25.1
google-auth==2.40.3
google-cloud-texttospeech==2.27.0
googleapis-common-protos==1.70.0
grpcio==1.73.1
grpcio-status==1.73.1
idna==3.10
proto-plus==1.26.1
protobuf==6.31.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
requests==2.32.4
rsa==4.9.1
tqdm==4.67.1
urllib3==2.5.0
